Determine if groups are numeric using is.numeric
As discussed in #5, the overlapMatrix function needs to determine if
the elements of the `group` list are meant to be treated like integer
indices or character indices. Currently, it tests if the class of the
first group is identically "numeric" but if you pass a list of
integers (as the documentation suggests), they will be of class
"integer" and this will fail, and overlapMatrix will proceed as if you
have passed characters.

If we test using `is.numeric(x)` rather than `class(x) == "numeric")`
we will be more robust.