Let grpreg handle group.multiplier construction
Currently, if the user specifies group.multiplier, it is passed along
to grpreg, which will use it as specified. However, if the user does
not specify a group.multiplier, grpregOverlap will construct it and
then pass this on to grpreg. While this enables additional
functionality (like dropping empty groups, or deleting variables that
do not appear in X), this prevents grpreg's default behavior, where
group sizes are based on the group-wise orthogonalized design matrix.
If there exist any groups with size g, and g>n, their effective size
will be reduced to (n-1), since g-many n-vectors can span at most R^n,
and due to centering it's R^(n-1). The "default" group.multiplier
value should be sqrt(n-1), which is what would be computed if grpreg
is called directly, but grpregOverlap will set the group.multiplier to
sqrt(g1), and then because this value is explicitly passed to grpreg,
it will not recompute the group.multipliers.

This problem is exacerbated by the way that cv.grpregOverlap works;
if the user does not specify group.multiplier, the "fit" object (fit
on the whole data) will use these "grpregOverlap"-style
group.multiplier defaults, whereas because cvfit results from calling
cv.grpreg with the latent representation, cross-validation will occur
using the grpreg-style group.multiplier defaults. If these
disagree (e.g., there exists a group with more than n-1 features),
then the results from cvfit and fit are not comparable, and lambda
tuning will correspond to a differently penalized model than what is
used in prediction and coefficient extraction.

This commit removes virtually all mention of group.multiplier from
grpregOverlap other than to specify that the `...` can be used to pass
options like `group.multiplier` through to grpreg.

Handling of problematic groups (e.g., those that contain features not
in X or are composed of only such features) should still work since
this is incorportated into incidenceMatrix, and will be reflected in grp.vec.