remove (almost) all arguments from grpregOverlap shared with grpreg
Many of the arguments that grpregOverlap takes are also arguments to
grpreg, and at present it manually passes these through. However,
there are a number of cases where the defaults differ (e.g., max.iters
is 1000 in grpregOverlap but 10,000 in grpreg) which may have resulted
from drift between the packages over time.

This commit removes almost all of these arguments from the signature
and the accompanying documentation for the
grpregOverlap::grpregOverlap function. Users can instead pass these
arguments using `...` but will need to refer to the grpreg package's
documentation to learn about them and their defaults.