migrate away from soft-deprecated expect_that testthat usage
Current tests use `not` for negation in testthat, but this is now
deprecated and causes warnings during testing. Indeed, the
`expect_that` construct has been soft-deprecated.

This commit replaces all `expect_that` expressions with comparable
expressions compatible with current `testthat` release.

The effect of some tests is slightly different. For example, it used
to expect that a specific error not be thrown, but if a different
error was thrown this would satisfy the test. Now, it simply expects
there to be *no* error, and the test fails if any errors are raised.